REFACTOR: KNSearch 모듈 분리
