+++ Projects/App/Sources/Presentation/Home Screen/HomeScreenFeature.swift
@@ -8,7 +8,6 @@
 import ComposableArchitecture
 import Factory
 import Foundation
-import KNCore
 import KNDomain
 import KNUtility
