+++ Projects/App/Sources/Presentation/ParentView/View/ParentViewController+Binding.swift
@@ -6,7 +6,6 @@
 //
 import Combine
 import Foundation
-import KNCore
 import KNUtility
 import UIKit
