+++ Projects/App/Sources/Presentation/TabBar/ViewModel/TabBarViewModel.swift
@@ -9,7 +9,6 @@ import Combine
 import CorePresentation
 import Factory
 import Foundation
-import KNCore
 import KNDeepLink
 import KNUtility
 import os
