+++ Projects/Features/KNSearch/Project.swift
@@ -1,0 +1,22 @@
+import ProjectDescription
+import ProjectDescriptionHelpers
+
+let project = Project.module(
+    name: "KNSearch",
+    dependencies: [
+        .project(target: "KNUtility", path: "../../KNUtility"),
+        .project(target: "KNDesignSystem", path: "../../KNDesignSystem"),
+        .project(target: "UIComponents", path: "../../UIComponents"),
+        .project(target: "CorePresentation", path: "../CorePresentation"),
+        .project(target: "KNDomain", path: "../../KNDomain"),
+        .project(target: "KNData", path: "../../KNData"),
+        .external(name: "Factory"),
+        .external(name: "RxSwift"),
+        .external(name: "RxCocoa"),
+        .external(name: "RxDataSources"),
+        .external(name: "SnapKit"),
+        .external(name: "FirebaseAnalytics"),
+        .external(name: "ComposableArchitecture"),
+    ],
+    hasTests: false
+)
