+++ Projects/Features/KNSearch/Sources/Interfaces/Searchable.swift
