+++ Projects/Features/KNSearch/Sources/View/SearchViewBackground.swift
